requirements.txt
Django==1.9.8
djangorestframework==3.4.1
djangorestframework-jwt==1.8.0
MySQL-python==1.2.5
Pillow==3.3.0
PyJWT==1.4.1
